refactor: delete class authorisation
